drivers: interrupt_controller: refactor GIC driver
GICv3 and GICv2 GICv1 drivers are refactored to use common functions
for distributor interface. Headers are reorgalized accordingly.
