+++ drivers/interrupt_controller/CMakeLists.txt
@@ -1,8 +1,9 @@
 # SPDX-License-Identifier: Apache-2.0
 
 zephyr_sources_ifdef(CONFIG_ARCV2_INTERRUPT_UNIT    intc_arcv2_irq_unit.c)
-zephyr_sources_ifdef(CONFIG_GIC_V1                  intc_gic.c)
-zephyr_sources_ifdef(CONFIG_GIC_V2                  intc_gic.c)
+zephyr_sources_ifdef(CONFIG_GIC                     intc_gic_common.c)
+zephyr_sources_ifdef(CONFIG_GIC_V1                  intc_gicv1v2.c)
+zephyr_sources_ifdef(CONFIG_GIC_V2                  intc_gicv1v2.c)
 zephyr_sources_ifdef(CONFIG_GIC_V3                  intc_gicv3.c)
 zephyr_sources_ifdef(CONFIG_IOAPIC                  intc_ioapic.c)
 zephyr_sources_ifdef(CONFIG_LOAPIC                  intc_loapic.c intc_system_apic.c)
